+++ ic-exports/src/pocket_ic.rs
@@ -9,7 +9,7 @@ use once_cell::sync::Lazy;
 pub use pocket_ic::*;
 
 #[cfg(feature = "pocket-ic-tests-async")]
-pub mod asnc;
+pub mod nio;
 
 const DFINITY_REVISION: &str = "307d5847c1d2fe1f5e19181c7d0fcec23f4658b3";
